+++ js/modules/main.js
@@ -1,0 +1,5 @@
+import { setupFormHandler } from "./formHandler";
+
+document.addEventListener("DOMContentLoaded", () => {
+    setupFormHandler();
+});
